Integrating MongoDB - Post Request
In order to add data to the DB (and also read from it), the async function saveURL is called on the post request, and the URL analysis is done there, with the assistance of Mongoose functions to read from the DB (instead of the urls array). The dns.lookup function was promisified in order to allow the correct flow of the async function.